src/emris_network_backend/task_manager_impl.rs
// Import necessary modules and types.
use crate::model_chunk::ModelChunk;
use crate::task_manager::TaskManagerInterface;
use crate::user::User;
use crate::completion::Completion;
use crate::training_task::TrainingTask;
use crate::webgpu_compute::run_gpu_computation;
use std::collections::HashMap;

// Define the TaskManagerImpl struct, which contains fields for managing users, model chunks,
// models, and training tasks.
pub struct TaskManagerImpl {
    users: HashMap<String, User>,
    model_chunks: HashMap<String, ModelChunk>,
    models: HashMap<String, Model>,
    training_tasks: HashMap<String, TrainingTask>,
}

// Define the Model struct, which contains fields for the model ID, minimum resources required,
// and whether the model is active.
#[derive(Clone, Deserialize, Serialize, CandidType)]
pub struct Model {
    pub id: String,
    pub min_resources: u64,
    pub active: bool,
}

// Implement the Default trait for TaskManagerImpl.
impl Default for TaskManagerImpl {
    fn default() -> Self {
        Self {
            users: HashMap::new(),
            model_chunks: HashMap::new(),
            models: HashMap::new(),
            training_tasks: HashMap::new(),
        }
    }
}

// Implement the TaskManagerInterface trait for TaskManagerImpl.
impl TaskManagerInterface for TaskManagerImpl {
    // Initialize the TaskManagerImpl.
    fn init(&mut self) {
        // TODO: Initialize any additional fields or state here.
        // Initialize total_resources
        // self.total_resources = 0;
    }

    // Register a new user.
    fn register_user(&mut self, user: User) -> Result<String, String> {
        // Check if the user already exists.
        if self.users.contains_key(&user.id) {
            return Err("User already exists.".to_string());
        }
        // Add the user to the users HashMap and return the user ID.
        let user_id = user.id.clone();
        self.users.insert(user.id, user);
        Ok(user_id)
    }

    // Update the resources allocated to a user.
    fn update_user_resources(&mut self, id: &str, resources: u64) -> Result<(), String> {
        // Find the user in the users HashMap and update their resources.
        if let Some(user) = self.users.get_mut(id) {
            user.resources = resources;
            Ok(())
        } else {
            Err("User not found.".to_string())
        }
    }

    // Distribute model chunks to users.
    fn distribute_model_chunks(&mut self) -> Result<(), String> {
        // TODO: Implement logic to distribute model chunks to users.
        // This is just a simple example.
        let chunk_id = "chunk1";
        let user_id = "user1";
        let data = vec![0u8; 1024];
        let chunk = ModelChunk {
            id: chunk_id.to_string(),
            user_id: user_id.to_string(),
            data: data,
        };
        self.model_chunks.insert(chunk_id.to_string(), chunk);
        Ok(())
    }

    // Submit a computed model chunk.
    fn submit_computed_chunk(&mut self, chunk: ModelChunk) -> Result<(), String> {
        // Check if the model chunk exists in the model_chunks HashMap.
        if let Some(existing_chunk) = self.model_chunks.get_mut(&chunk.id) {
            // Update the data of the existing model chunk.
            existing_chunk.data = chunk.data;
        } else {
            return Err("Model chunk not found.".to_string());
        }

        // Calculate and update user rewards.
           let reward = self.calculate_rewards(&chunk.user_id, 1);
    println!("User {} earned {} tokens.", chunk.user_id, reward);

    Ok(())
}

// Get model chunks associated with a user.
fn get_model_chunks(&self, user_id: &str) -> Result<Vec<ModelChunk>, String> {
    let mut chunks = Vec::new();
    // Iterate over all model chunks and find the ones associated with the user.
    for chunk in self.model_chunks.values() {
        if chunk.user_id == user_id {
            chunks.push(chunk.clone());
        }
    }
    Ok(chunks)
}

// Get the rewards earned by a user.
fn get_rewards(&self, user_id: &str) -> Result<u64, String> {
    // Find the user in the users HashMap and return their rewards.
    if let Some(user) = self.users.get(user_id) {
        Ok(user.rewards)
    } else {
        Err("User not found.".to_string())
    }
}

// Register a new model.
fn register_model(&mut self, model: Model) -> Result<String, String> {
    // Check if the model already exists.
    if self.models.contains_key(&model.id) {
        return Err("Model already exists.".to_string());
    }
    // Add the model to the models HashMap and return the model ID.
    let model_id = model.id.clone();
    self.models.insert(model.id, model);
    Ok(model_id)
}

// Activate a model.
fn activate_model(&mut self, model_id: &str) -> Result<(), String> {
    // Calculate the total resources available.
    let total_resources = self.users.values().map(|user| user.resources).sum();
    // Find the model in the models HashMap.
    let model = self.models.get_mut(model_id).ok_or("Model not found.")?;
    // Check if there are sufficient resources to activate the model.
    if total_resources >= model.min_resources {
        model.active = true;
        Ok(())
    } else {
        Err("Insufficient resources to activate the model.".to_string())
    }
}

// Deactivate a model.
fn deactivate_model(&mut self, model_id: &str) -> Result<(), String> {
    // Find the model in the models HashMap and deactivate it.
    let model = self.models.get_mut(model_id).ok_or("Model not found.")?;
    model.active = false;
    Ok(())
}

// Get all active models.
fn get_active_models(&self) -> Vec<Model> {
    self.models.values().filter(|model| model.active).cloned().collect()
}

// Get models that need more resources to be activated.
fn get_models_needing_resources(&self) -> Vec<Model> {
    let total_resources = self.users.values().map(|user| user.resources).sum();
    self.models.values()
        .filter(|model| !model.active && model.min_resources > total_resources)
        .cloned()
        .collect()
}

// Submit a computed model chunk and process it using WebGPU.
fn submit_computed_chunk(&mut self, chunk: ModelChunk) -> Result<(), String> {
    // Use WebGPU to process the model chunk data.
    let input_data = chunk.data;
    let result = match block_on(run_gpu_computation(input_data)) {
        Ok(result) => result,
        Err(err) => return Err(err),
    };

    // TODO: Use the result for further processing (e.g., update rewards, update model).
    // ...

    Ok(())
}

// Generate a completion based on a given prompt.
fn generate_completion(&self, prompt: &str) -> Result<Completion, String> {
    // Get all active models.
    let active_models = self.get_active_models();
    // Check if there are any active models available.
    if active_models.is_empty() {
        return Err("No active models available.".to_string());
    }

    // Use the first active model (for simplicity).
    let model = &active_models[0];
    // TODO: Implement logic for distributed model processing.
    // Simulate distributed model processing.
    let generated_text = format!(
        "Generated text for '{}' using model '{}'",
        prompt, model.id
    );

    // Return the generated completion.
    Ok(Completion {
        prompt: prompt.to_string(),
        generated_text,
    })
}

// Create a new training task.
fn create_training_task(&mut self, task: TrainingTask) -> Result<String, String> {
    // Check if the training task already exists.
    if self.training_tasks.contains_key(&task.id) {
        return Err("Training task already exists.".to_string());
    }
    // Add the training task to the training_tasks HashMap and return the task ID.
    let task_id = task.id.clone();
    self.training_tasks.insert(task.id, task);
    Ok(task_id)
}

// Submit training results for a training task.
fn submit_training_results(
    &mut self,
    task_id: &str,
    model_weights: Vec<u8>,
) -> Result<(), String> {
    // Find the training task in the training_tasks HashMap and update its model weights.
    if let Some(task) = self.training_tasks.get_mut(task_id) {
        task.model_weights = Some(model_weights);
        Ok(())
    } else {
        Err("Training task not found.".to_string())
    }
}
}

// Implement additional methods for TaskManagerImpl.
impl TaskManagerImpl {
// Calculate rewards for a user based on the number of completed chunks.
fn calculate_rewards(&mut self, user_id: &str, completed_chunks: usize) -> u64 {
// TODO: Implement reward calculation logic here.
// For example, you could reward users with 1 token per completed chunk.
let reward = completed_chunks as u64;
// Update the user's rewards in the users HashMap.
if let Some(user) = self.users.get_mut(user_id) {
user.rewards += reward;
}
reward
}
}
